logger: widen the session byte counter, and prove a split part opens on its own
Two things found by checking the overnight path properly rather than reasoning
about it.

g_status.bytes was uint32. Both IMUs at the native rate write 227 KB/s, which
passes the 4 GB ceiling after about 5.3 hours - so an overnight run would have
reported a wrapped total, which is exactly the number you look at in the
morning to decide whether it went well. It is now uint64. The data was never at
risk: the rollover uses the per-part counter, which is capped at 100 MB.

The split is now tested rather than argued. tests/ulog_split_test.py builds
parts with logger.c's exact byte layout - different start timestamps, different
data - and requires pyulog to open each one alone. It also strips the prologue
from a third file and requires that one to FAIL, so the test cannot pass
vacuously if the rollover ever stops writing a header. That was the real risk:
a missing per-part prologue leaves 100 MB of readable data and gigabytes of
rubble, with nothing on the board reporting anything wrong.

Also verified while looking: sscanf("log_007_02.ulg", "log_%d") really does
yield 7, so session numbering does not collide with split parts and cannot
overwrite an existing log; the daemon completes every log_data() before the
rollover check, so no record is torn across parts; and g_buflen is size_t, so
the 64 KB buffer does not wrap an index.

+++ apps/logger/logger.h
@@ -39,7 +39,14 @@ struct logger_status_s
   char     path[48];       /* the .ulg being written */
   uint32_t topics;         /* how many topics are in this session */
   uint32_t samples;        /* records written */
-  uint32_t bytes;          /* bytes written to the card */
+
+  /* 64-bit, because 32 is not enough. Both IMUs at the native 2 kHz write
+   * 227 KB/s, which passes 4 GB - the uint32 ceiling - after about 5.3 hours.
+   * An overnight run would have reported a wrapped total, which is precisely
+   * the number you check in the morning to see whether it went well.
+   */
+
+  uint64_t bytes;          /* bytes written across the whole session */
   uint32_t dropped;        /* samples dropped because the card fell behind */
   int32_t  rate;           /* LOG_RATE this session was started with */
 };
